Desktop version of event page hero

+++ client/src/common/components/category/Category.tsx
@@ -7,26 +7,31 @@ import styles from "./category.module.scss";
 
 interface Props {
   category: CategoryType;
-  onRemove: (category: CategoryType) => void;
+  onRemove?: (category: CategoryType) => void;
 }
 
 const Category: FunctionComponent<Props> = ({ category, onRemove }) => {
   const handleRemove = () => {
-    onRemove(category);
+    if (onRemove) {
+      onRemove(category);
+    }
   };
 
   return (
     <div className={styles.category}>
-      <button
-        type="button"
-        className={styles.closeButton}
-        onClick={handleRemove}
-        aria-label={formatMessage("commons.category.ariaButtonRemove", {
-          category: category.text
-        })}
-      >
-        <CloseIcon />
-      </button>
+      {onRemove && (
+        <button
+          type="button"
+          className={styles.closeButton}
+          onClick={handleRemove}
+          aria-label={formatMessage("commons.category.ariaButtonRemove", {
+            category: category.text
+          })}
+        >
+          <CloseIcon />
+        </button>
+      )}
+
       {category.text}
     </div>
   );
